requirements.txt
jax[cuda12]
# jax[cpu] on a machine without a GPU
gymnax
optax
distrax
flax
flashbax
hydra-core
joblib